+++ index.js
@@ -127,8 +127,8 @@ game.control_map_set_default("CursorMovementY",
         "Mouse_MovementY_Positive"
     ]
 );
-game.control_map_set_default("TouchX", "Touch_ClientX_Positive");
-game.control_map_set_default("TouchY", "Touch_ClientY_Positive");
+game.control_map_set_default("TouchX", "Touch_MovementX_Positive");
+game.control_map_set_default("TouchY", "Touch_MovementY_Positive");
 game.control_map_restore_default();
 game.control_map_save();
 
@@ -256,16 +256,21 @@ player.update_function = function (engine) {
         movementY = engine.control_map_get("CursorMovementY", false),
         clientX = engine.control_map_get("CursorX", true),
         clientY = engine.control_map_get("CursorY", true),
-        touchX = engine.control_map_get("TouchX"),
-        touchY = engine.control_map_get("TouchY");
+        touchX = engine.control_map_get("TouchX", true),
+        touchY = engine.control_map_get("TouchY", true);
     camera_hud.width = engine.width;
     camera_hud.height = engine.height;
-    this.cursor_update(engine, movementX, movementY, true);
     if (touchX && touchY) {
-        if (touchX[0].state === "Down" && touchY[0].state === "Down") {
-            this.cursor_update(engine, touchX[0].value, touchY[0].value, false);
+        for (let i = 0; i < touchX.length; i++) {
+            if (touchX[i].state === "Down" && touchY[i].state === "Down") {
+                movementX = touchX[i].value;
+                touchX[i].value = 0;
+                movementY = touchY[i].value;
+                touchY[i].value = 0;
+            }
         }
     }
+    this.cursor_update(engine, movementX, movementY, true);
     this.game_title.element[0].string =
         engine.title + "\n" +
         eated_walls + "/" + total_walls + " Eated Walls";